+++ core/src/test/resources/specmaticConfigFiles/specmatic_config_v1.json
@@ -4,6 +4,7 @@
     {
       "type": "git",
       "repository": "https://contracts",
+      "branch": "1.0.1",
       "provides": [
         "com/petstore/1.yaml"
       ],
