+++ core/src/test/resources/specmaticConfigFiles/specmatic_config_v1.yaml
@@ -2,6 +2,7 @@ version: 1
 contract_repositories:
   - type: git
     repository: https://contracts
+    branch: 1.0.1
     provides:
       - com/petstore/1.yaml
     consumes:
