+++ core/src/test/resources/specmaticConfigFiles/specmatic_without_version.json
@@ -3,6 +3,7 @@
     {
       "type": "git",
       "repository": "https://contracts",
+      "branch": "1.0.1",
       "provides": [
         "com/petstore/1.yaml"
       ],
