+++ core/src/test/resources/specmaticConfigFiles/specmatic_without_version.yaml
@@ -1,6 +1,7 @@
 contract_repositories:
   - type: git
     repository: https://contracts
+    branch: 1.0.1
     provides:
       - com/petstore/1.yaml
     consumes:
